Move the dange-zone to its own modal
